+++ Fabricor/VulkanRendering/VoxelRenderer/VoxelMesh.cs
@@ -6,12 +6,12 @@ namespace Fabricor.VulkanRendering.VoxelRenderer
 {
     public class VoxelMesh{
         public FDataBuffer<VoxelVertex> vertices;
-        public FDataBuffer<ushort> indices;
+        public FDataBuffer<uint> indices;
 
-        public VoxelMesh(VkDevice device,VkPhysicalDevice physicalDevice, VoxelVertex[] vertices,ushort[] indices){
+        public VoxelMesh(VkDevice device,VkPhysicalDevice physicalDevice, VoxelVertex[] vertices,uint[] indices){
             this.vertices=new FDataBuffer<VoxelVertex>(device,physicalDevice,vertices.Length,VkBufferUsageFlags.VertexBuffer,
             VkSharingMode.Exclusive);
-            this.indices=new FDataBuffer<ushort>(device,physicalDevice,indices.Length,VkBufferUsageFlags.IndexBuffer,VkSharingMode.Exclusive);
+            this.indices=new FDataBuffer<uint>(device,physicalDevice,indices.Length,VkBufferUsageFlags.IndexBuffer,VkSharingMode.Exclusive);
 
             Span<VoxelVertex> spanv=this.vertices.Map();
             for (int i = 0; i < vertices.Length; i++)
@@ -20,7 +20,7 @@ namespace Fabricor.VulkanRendering.VoxelRenderer
             }
             spanv=this.vertices.UnMap();
 
-            Span<ushort> spani=this.indices.Map();
+            Span<uint> spani=this.indices.Map();
             for (int i = 0; i < indices.Length; i++)
             {
                 spani[i]=indices[i];
